+++ src/main/java/com/yzpc/yzpc_weixinapp/service/ApplicationService.java
@@ -14,7 +14,7 @@ public interface ApplicationService extends IService<Application> {
 
     void addApplication(Application application);
 
-    List<Application> getApplicationsById(Integer studentId);
+    List<Application> getApplicationsById(Long studentId);
 
     Application getApplicationsByOwnId(Long applicationId);
